backend/media_protection/requirements.txt
transformers
torch>=2.0.0
diffusers
accelerate
scipy
safetensors
pydantic
requests
pillow
torchvision
matplotlib
tqdm
